fix(catalog): isolate explicit video capability profiles
